Add in server side validation

+++ server.js
@@ -6,12 +6,13 @@ const bodyparser = require('body-parser');
 const socketIO = require('socket.io');
 const nodeController = require('./controller');
 const Node = require('./dbmodel/node');
+const Joi = require('@hapi/joi');
 
 const port = process.env.PORT || 3000;
 
 const app = express();
 
-mongoose.connect('mongodb://tree:sockettree1@ds115035.mlab.com:15035/tree')
+mongoose.connect('mongodb://tree:sockettree1@ds115035.mlab.com:15035/tree', { useFindAndModify: false, useNewUrlParser: true })
 .then(() => {
     console.log('Connected');
 });
@@ -19,6 +20,14 @@ mongoose.connect('mongodb://tree:sockettree1@ds115035.mlab.com:15035/tree')
 app.use(bodyparser.urlencoded({ extended: false }));
 app.use(bodyparser.json());
 app.use(express.static(path.join(__dirname, '/dist/tree')));
+
+const schema = Joi.object().keys({
+    id: Joi.string().alphanum(),
+    name: Joi.string(),
+    count: Joi.number().max(15),
+    lower: Joi.number(),
+    higher: Joi.number().greater(Joi.ref('lower'))
+})
 
 const server = http.createServer(app);
 const io = socketIO(server);
@@ -28,16 +37,39 @@ app.set('io', io);
 io.on('connection', (socket) => {
     console.log("Connected to Socket!!"+ socket.id);
     socket.on('addNode', (node) => {
-        nodeController.addNode(io,node);
+        Joi.validate(node, schema, function (err, value) {
+            if(err){
+                console.log(err.message);
+            }else{
+                nodeController.addNode(io,node);
+            }
+        });
     });
     socket.on('renameNode', (node) => {
-        nodeController.renameNode(io,node);
+        Joi.validate(node, schema, function (err, value) {
+            if(err){
+                console.log(err.message);
+            }else{
+                nodeController.renameNode(io,node);
+            }
+        });
     });
     socket.on('regenerateChild', (node) => {
-        nodeController.regenerateChild(io,node);
+        debugger;
+        Joi.validate(node, schema, function (err, value) {
+            if(err){
+                console.log(err.message);
+            }else{
+                nodeController.regenerateChild(io,node);
+            }
+        });
     });
     socket.on('deleteNode', (id) => {
-        nodeController.deleteNode(io,id);
+        if(!id){
+            console.log("Error: id is incorrect - " + id); 
+        }else{
+            nodeController.deleteNode(io,id);
+        }
     });
 });
